test: split disallowed AST pattern tests to satisfy file line gate
repo.dart_file_non_comment_line_size failed on check_disallowed_ast_patterns_test.dart after new cases; extract shared YAML fixture and move prohibited_linear_units_armies_fleets_lookup group to its own test file.

Refs #2394
